+++ tests/unit_tests/CMakeLists.txt
@@ -97,6 +97,7 @@ set(unit_tests_sources
   test_protocol_pack.cpp
   threadpool.cpp
   tree_cache.cpp
+  tx_construction_helpers.cpp
   tx_proof.cpp
   hardfork.cpp
   unbound.cpp
