+++ tests/unit_tests/fake_pruned_blockchain.h
@@ -30,75 +30,11 @@
 
 #define IN_UNIT_TESTS
 
-#include "carrot_impl/carrot_tx_builder_types.h"
 #include "curve_trees.h"
 #include "wallet/wallet2.h"
 
 namespace mock
 {
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-bool construct_miner_tx_fake_reward_1out(const size_t height,
-    const rct::xmr_amount reward,
-    const cryptonote::account_public_address &miner_address,
-    cryptonote::transaction& tx,
-    const uint8_t hf_version,
-    const size_t num_tx_outputs = 1);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-cryptonote::transaction construct_miner_tx_fake_reward_1out(const size_t height,
-    const rct::xmr_amount reward,
-    const cryptonote::account_public_address &miner_address,
-    const uint8_t hf_version,
-    const size_t num_tx_outputs = 1);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-struct stripped_down_tx_source_entry_t
-{
-    uint64_t global_output_index;
-    crypto::public_key onetime_address;
-    crypto::public_key real_out_tx_key;
-    std::vector<crypto::public_key> real_out_additional_tx_keys;
-    size_t local_output_index;
-    rct::xmr_amount amount;
-    rct::key mask;
-};
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-cryptonote::tx_source_entry gen_tx_source_entry_fake_members(
-    const stripped_down_tx_source_entry_t &in,
-    const size_t mixin,
-    const uint64_t max_global_output_index);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-cryptonote::transaction construct_pre_carrot_tx_with_fake_inputs(
-    const cryptonote::account_keys &sender_account_keys,
-    const std::unordered_map<crypto::public_key, cryptonote::subaddress_index> &sender_subaddress_map,
-    std::vector<stripped_down_tx_source_entry_t> &&stripped_sources,
-    std::vector<cryptonote::tx_destination_entry> &destinations,
-    const boost::optional<cryptonote::account_public_address> &change_addr,
-    const rct::xmr_amount fee,
-    const uint8_t hf_version,
-    const bool sweep_unmixable_override = false);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-cryptonote::transaction construct_pre_carrot_tx_with_fake_inputs(
-    std::vector<cryptonote::tx_destination_entry> &destinations,
-    const rct::xmr_amount fee,
-    const uint8_t hf_version,
-    const bool sweep_unmixable_override = false);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-static constexpr rct::xmr_amount fake_fee_per_weight = 2023;
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-cryptonote::transaction construct_carrot_pruned_transaction_fake_inputs(
-    const std::vector<carrot::CarrotPaymentProposalV1> &normal_payment_proposals,
-    const std::vector<carrot::CarrotPaymentProposalVerifiableSelfSendV1> &selfsend_payment_proposals,
-    const cryptonote::account_keys &acc_keys);
-//----------------------------------------------------------------------------------------------------------------------
-//----------------------------------------------------------------------------------------------------------------------
-extern const cryptonote::account_public_address null_addr;
 //----------------------------------------------------------------------------------------------------------------------
 //----------------------------------------------------------------------------------------------------------------------
 using fcmp_generic_object_t = std::unique_ptr<void, decltype(&free)>;
